下单模块开发, 区分普通下单和策略下单。
