+++ DEV_LOG.md
@@ -120,3 +120,9 @@ npm install --legacy-peer-deps
 对于策略的调度，可以考虑使用消息队列来进行解耦，使得策略中心和调度器之间可以进行通信。（GPT写的，但是我觉得不错就mark一下）
 
 但是需要考虑的一个问题是在开单时，如何保证余额的正确性，以及如何保证交易的完整性。
+
+目前下单分为普通下单和策略下单，这两个的职能是不同的：
+
+普通下单的目的是完成交易：以多少的价格购买多少数量的某币种。
+
+策略交易的目的是添加止盈止损等功能：如提供单向止盈止损委托、双向止盈止损委托、计划委托、时间加权委托、移动止盈止损委托等。
