+++ backend/service/post_order_service.py
@@ -1,6 +1,5 @@
 import okx.Trade as Trade
-from backend.data_center.data_object.enum_obj import *
-from backend.data_center.data_object.req.post_order_req import PostOrderReq
+from backend.data_center.data_object.req.place_order.place_order_req import PostOrderReq
 from backend.service.utils import *
 
 
@@ -82,8 +81,8 @@ class TradeAPI:
             return {"status": "error", "error_message": str(order_exception)}
 
     # 下单
-    def get_order_info(self, type: str, instId: str, orderId: str):
-        if type.__eq__(EnumTradeType.DEMO.value):
+    def get_order_info(self, env: str, instId: str, orderId: str):
+        if env.__eq__(EnumTradeType.DEMO.value):
             okx = self.get_okx_trade_demo()
         else:
             okx = self.get_okx_trade_product()
